Add message, thread entity & Separate YoUPS schema into different file
