+++ schema/models.py
@@ -8,6 +8,8 @@ from oauth2client.django_orm import FlowField, CredentialsField
 from http_handler import settings
 from http_handler.settings import AUTH_USER_MODEL
 
+from schema.youps import ImapAccount
+
 class Post(models.Model):
 	id = models.AutoField(primary_key=True)
 	author = models.ForeignKey(settings.AUTH_USER_MODEL, related_name='user_authored_posts', null=True)
@@ -321,48 +323,6 @@ class UserProfile(AbstractBaseUser):
 	def is_staff(self):
 		"Is the user a member of staff?"
 		return self.is_admin
- 
-class ImapAccount(models.Model):
-	id = models.AutoField(primary_key=True)
-	newest_msg_id = models.IntegerField(default=-1)
-
-	email = models.EmailField(
-        verbose_name='email address',
-        max_length=255,
-        unique=True,
-    )
-	password = models.CharField('password', max_length=100, blank=True)
-	host = models.CharField('host', max_length=100)
-
-	is_oauth = models.BooleanField(default=False)
-	access_token = models.CharField('access_token', max_length=200, blank=True)
-	refresh_token = models.CharField('refresh_token', max_length=200, blank=True)
-
-	current_mode = models.ForeignKey('MailbotMode', null=True, blank=True)
-	shortcuts = models.TextField(default="")
-
-	# code = models.TextField(null=True, blank=True)
-	execution_log = models.TextField(default="")
-	is_test = models.BooleanField(default=True)
-	is_running = models.BooleanField(default=False)
-
-	arrive_action = models.CharField('access_token', max_length=1000, blank=True)
-	custom_action = models.CharField('custom_action', max_length=1000, blank=True)
-	timer_action = models.CharField('timer_action', max_length=1000, blank=True)
-	repeat_action = models.CharField('repeat_action', max_length=1000, blank=True)
-
-	# user = models.ForeignKey(settings.AUTH_USER_MODEL, null=True, blank=True)
-
-class MailbotMode(models.Model):
-	uid = models.IntegerField()
-
-	name = models.CharField('mode_name', max_length=100)
-	code = models.TextField(null=True, blank=True)
-
-	imap_account = models.ForeignKey('ImapAccount')
-
-	class Meta:
-		unique_together = ("uid", "imap_account")
 
 class Following(models.Model):
 	id = models.AutoField(primary_key=True)
